+++ src/main/java/org/gridsuite/odre/server/client/OdreCsvClient.java
@@ -1,0 +1,24 @@
+/**
+ * Copyright (c) 2019, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package org.gridsuite.odre.server.client;
+
+import org.gridsuite.odre.server.dto.LineGeoData;
+import org.gridsuite.odre.server.dto.SubstationGeoData;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.util.List;
+import java.util.Map;
+
+/**
+ * @author bendaamerahm <ahmed.bendaamer at rte-france.com>
+ */
+public interface OdreCsvClient {
+
+    List<SubstationGeoData> getSubstationsFromCsv(MultipartFile file);
+
+    List<LineGeoData> getLinesFromCsv(Map<String, MultipartFile> files);
+}
