+++ src/main/java/org/gridsuite/odre/server/services/OdreServiceImpl.java
@@ -6,11 +6,10 @@
  */
 package org.gridsuite.odre.server.services;
 
-import org.gridsuite.odre.server.utils.FileNameEnum;
 import org.gridsuite.odre.server.client.OdreClient;
+import org.gridsuite.odre.server.client.OdreCsvClient;
 import org.gridsuite.odre.server.dto.LineGeoData;
 import org.gridsuite.odre.server.dto.SubstationGeoData;
-import org.gridsuite.odre.server.utils.GeographicDataParser;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
@@ -25,11 +24,6 @@ import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.util.DefaultUriBuilderFactory;
 import org.springframework.web.util.UriComponentsBuilder;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.UncheckedIOException;
-import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -42,11 +36,13 @@ import java.util.Objects;
 public class OdreServiceImpl implements OdreService {
 
     private static final String GEO_DATA_API_VERSION = "v1";
-    public static final String UTF_8 = "UTF-8";
 
     @Autowired
     @Qualifier("odreDownloadClientImpl")
     private OdreClient client;
+
+    @Autowired
+    private OdreCsvClient csvClient;
 
     private RestTemplate geoDataServerRest;
 
@@ -98,7 +94,7 @@ public class OdreServiceImpl implements OdreService {
         HttpHeaders requestHeaders = new HttpHeaders();
         requestHeaders.setContentType(MediaType.APPLICATION_JSON);
         UriComponentsBuilder uriBuilder = UriComponentsBuilder.fromHttpUrl(geoDataServerBaseUri + "/" + GEO_DATA_API_VERSION + "/substations");
-        List<SubstationGeoData> substationsGeoData = new ArrayList<>(getSubstationsFromCsv(file));
+        List<SubstationGeoData> substationsGeoData = new ArrayList<>(csvClient.getSubstationsFromCsv(file));
         HttpEntity<List<SubstationGeoData>> requestEntity = new HttpEntity<>(substationsGeoData, requestHeaders);
         geoDataServerRest.exchange(uriBuilder.toUriString(),
                 HttpMethod.POST,
@@ -111,32 +107,11 @@ public class OdreServiceImpl implements OdreService {
         HttpHeaders requestHeaders = new HttpHeaders();
         requestHeaders.setContentType(MediaType.APPLICATION_JSON);
         UriComponentsBuilder uriBuilder = UriComponentsBuilder.fromHttpUrl(geoDataServerBaseUri + "/" + GEO_DATA_API_VERSION + "/lines");
-        List<LineGeoData> linesFromCsv = getLinesFromCsv(files);
+        List<LineGeoData> linesFromCsv = csvClient.getLinesFromCsv(files);
         HttpEntity<List<LineGeoData>> requestEntity = new HttpEntity<>(linesFromCsv, requestHeaders);
         geoDataServerRest.exchange(uriBuilder.toUriString(),
                 HttpMethod.POST,
                 requestEntity,
                 Void.class);
     }
-
-    public List<SubstationGeoData> getSubstationsFromCsv(MultipartFile file) {
-        try (BufferedReader fileReader = new BufferedReader(new InputStreamReader(file.getInputStream(), StandardCharsets.UTF_8))) {
-            return new ArrayList<>(GeographicDataParser.parseSubstations(fileReader).values());
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
-        }
-    }
-
-    public List<LineGeoData> getLinesFromCsv(Map<String, MultipartFile> files) {
-        try (BufferedReader aerialBufferedReader = new BufferedReader(new InputStreamReader(files.get(FileNameEnum.AERIAL_LINES.getValue()).getInputStream(), StandardCharsets.UTF_8));
-             BufferedReader undergroundBufferedReader = new BufferedReader(new InputStreamReader(files.get(FileNameEnum.UNDERGROUND_LINES.getValue()).getInputStream(), StandardCharsets.UTF_8));
-             BufferedReader substationBufferedReader = new BufferedReader(new InputStreamReader(files.get(FileNameEnum.SUBSTATIONS.getValue()).getInputStream(), StandardCharsets.UTF_8));
-        ) {
-            return new ArrayList<>(GeographicDataParser.parseLines(aerialBufferedReader, undergroundBufferedReader,
-                    GeographicDataParser.parseSubstations(substationBufferedReader)).values());
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
-        }
-    }
-
 }
